+++ libxmp/src/xmp_gasnet_coarray.c
@@ -11,10 +11,10 @@
 //  { XMP_GASNET_LOCKHANDOFF,    _xmp_gasnet_lockhandoff }
 //};
 static gasnet_seginfo_t *_test_seginfo;
-static unsigned long _xmp_coarray_shift = 0;
+static unsigned long long _xmp_coarray_shift = 0;
 static void **_xmp_gasnet_buf;
 
-void _XMP_gasnet_set_coarray(_XMP_coarray_t *coarray, void **addr, long number_of_elements, size_t type_size){
+void _XMP_gasnet_set_coarray(_XMP_coarray_t *coarray, void **addr, unsigned long long number_of_elements, size_t type_size){
   int numprocs;
   void **each_addr;   // head address of a local array on each node
 
@@ -41,7 +41,7 @@ static void *_gasnet_getseg(gasnet_node_t node) {
   return _test_seginfo[node].addr;
 }
 
-void _XMP_gasnet_initialize(int argc, char **argv, int malloc_size){
+void _XMP_gasnet_initialize(int argc, char **argv, unsigned long long malloc_size){
   int numprocs;
 
   gasnet_init(&argc, &argv);
@@ -74,13 +74,15 @@ void _XMP_gasnet_sync_all(){
   GASNET_BARRIER();
 }
 
-void _XMP_gasnet_put(int dest_node, _XMP_coarray_t* dest, int dest_point, void *src_ptr, int src_point, int length){
+void _XMP_gasnet_put(int dest_node, _XMP_coarray_t* dest, unsigned long long dest_point, void *src_ptr, 
+										 unsigned long long src_point, unsigned long long length){
   dest_point *= dest->type_size;
   src_point  *= dest->type_size;
   gasnet_put_nbi_bulk(dest_node, dest->addr[dest_node]+dest_point, (char *)(src_ptr)+src_point, dest->type_size*length);
 }
 
-void _XMP_gasnet_get(void *dest_ptr, int dest_point, int src_node, _XMP_coarray_t *src, int src_point, int length){
+void _XMP_gasnet_get(void *dest_ptr, unsigned long long dest_point, int src_node, _XMP_coarray_t *src, 
+										 unsigned long long src_point, unsigned long long length){
   dest_point *= src->type_size;
   src_point  *= src->type_size;
   gasnet_get_bulk((char *)(dest_ptr)+dest_point, src_node, src->addr[src_node]+src_point, src->type_size*length);
